requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
Pillow
pymupdf
openai
requests
python-dotenv